# Conflicts:
#   README.md

+++ rok4generation/main/lib/JOINCACHE/Shell.pm
@@ -169,10 +169,16 @@ OverlayNtiff () {
     local config=$1
     local inTemplate=$2
 
-    overlayNtiff -f ${ONT_CONF_DIR}/$config ${OVERLAYNTIFF_OPTIONS}
-    if [ $? != 0 ] ; then echo $0 : Erreur a la ligne $(( $LINENO - 1)) >&2 ; exit 1; fi
-    rm -f ${TMP_DIR}/$inTemplate
-    rm -f ${ONT_CONF_DIR}/$config
+    if [[ "${work}" == "0" ]]; then
+        return
+    fi
+
+    if [ -f ${MNT_CONF_DIR}/$config ]; then
+        overlayNtiff -f ${ONT_CONF_DIR}/$config ${OVERLAYNTIFF_OPTIONS}
+        if [ $? != 0 ] ; then echo $0 : Erreur a la ligne $(( $LINENO - 1)) >&2 ; exit 1; fi
+        rm -f ${TMP_DIR}/$inTemplate
+        rm -f ${ONT_CONF_DIR}/$config
+    fi
 }
 
 ONTFUNCTION
@@ -181,6 +187,10 @@ my $CEPH_STORAGE_FUNCTIONS = <<'STORAGEFUNCTIONS';
 LinkSlab () {
     local target=$1
     local link=$2
+
+    if [[ "${work}" == "0" ]]; then
+        return
+    fi
 
     # On retire le pool des entrées
     target=`echo -n "$target" | sed "s#${PYR_POOL}/##"`
@@ -196,6 +206,19 @@ PushSlab () {
     local workMskName=$3
     local mskName=$4
 
+    if [[ "${work}" = "0" ]]; then
+        # On regarde si l'image à pousser est la dernière traitée lors d'une exécution précédente
+        if [[ "${imgName}" == "${last_slab}" ]]; then
+            echo "Last generated image slab found, now we work"
+            work=1
+        elif [[ ! -z $mskName && "${mskName}" == "${last_slab}" ]] ; then
+            echo "Last generated mask slab found, now we work"
+            work=1
+        fi
+
+        return
+    fi
+
     work2cache ${TMP_DIR}/$workImgName ${WORK2CACHE_IMAGE_OPTIONS} -pool ${PYR_POOL} $imgName
     if [ $? != 0 ] ; then echo $0 : Erreur a la ligne $(( $LINENO - 1)) >&2 ; exit 1; fi
     rm -f ${TMP_DIR}/$workImgName
@@ -205,11 +228,17 @@ PushSlab () {
         if [ $? != 0 ] ; then echo $0 : Erreur a la ligne $(( $LINENO - 1)) >&2 ; exit 1; fi
         rm -f ${TMP_DIR}/$workMskName
     fi
+
+    print_prog
 }
 
 PullSlab () {
     local input=$1
     local output=$2
+
+    if [[ "${work}" == "0" ]]; then
+        return
+    fi
 
     # On retire le pool du input
     input=`echo -n "$input" | sed "s#${PYR_POOL}/##"`
@@ -225,6 +254,10 @@ LinkSlab () {
     local target=$1
     local link=$2
 
+    if [[ "${work}" == "0" ]]; then
+        return
+    fi
+
     mkdir -p $(dirname $link)
     if [ $? != 0 ] ; then echo $0 : Erreur a la ligne $(( $LINENO - 1)) >&2 ; exit 1; fi
 
@@ -238,6 +271,19 @@ PushSlab () {
     local imgName=$2
     local workMskName=$3
     local mskName=$4
+
+    if [[ "${work}" = "0" ]]; then
+        # On regarde si l'image à pousser est la dernière traitée lors d'une exécution précédente
+        if [[ "${imgName}" == "${last_slab}" ]]; then
+            echo "Last generated image slab found, now we work"
+            work=1
+        elif [[ ! -z $mskName && "${mskName}" == "${last_slab}" ]] ; then
+            echo "Last generated mask slab found, now we work"
+            work=1
+        fi
+
+        return
+    fi
 
     local dir=`dirname ${PYR_DIR}/$imgName`
 
@@ -259,11 +305,17 @@ PushSlab () {
         if [ $? != 0 ] ; then echo $0 : Erreur a la ligne $(( $LINENO - 1)) >&2 ; exit 1; fi
         rm -f ${TMP_DIR}/$workMskName
     fi
+
+    print_prog
 }
 
 PullSlab () {
     local input=$1
     local output=$2
+
+    if [[ "${work}" == "0" ]]; then
+        return
+    fi
 
     cache2work -c zip $input ${TMP_DIR}/$output
     if [ $? != 0 ] ; then echo $0 : Erreur a la ligne $(( $LINENO - 1)) >&2 ; exit 1; fi
@@ -391,6 +443,34 @@ sub getMainScript {
 #                                   Group: Export function                                         #
 ####################################################################################################
 
+my $WORKANDPROG = <<'WORKANDPROG';
+progression=-1
+progression_file="$0.prog"
+lines_count=$(wc -l $0 | cut -d' ' -f1)
+start_line=0
+
+print_prog () {
+    tmp=$(( (${BASH_LINENO[-2]} - $start_line) * 100 / (${lines_count} - $start_line) ))
+    if [[ "$tmp" != "$progression" ]]; then
+        progression=$tmp
+        echo "$tmp" >$progression_file
+    fi
+}
+
+work=1
+
+# Test d'existence de la liste temporaire
+if [[ -f "${TMP_LIST_FILE}" ]] ; then 
+    # La liste existe, ce qui suggère que le script a déjà commencé à tourner
+    # On prend la dernière ligne pour connaître la dernière dalle complètement traitée
+    
+    last_slab=$(tail -n 1 ${TMP_LIST_FILE} | sed "s#^0/##")
+    echo "Script ${SCRIPT_ID} recall, work from slab ${last_slab}"
+    work=0
+fi
+
+WORKANDPROG
+
 =begin nd
 Function: getScriptInitialization
 
@@ -405,8 +485,10 @@ sub getScriptInitialization {
 
     # Variables
 
+    my $string = $WORKANDPROG;
+
     # On a précisé une méthode de fusion, on est dans le cas d'un JOINCACHE, on exporte la fonction overlayNtiff
-    my $string = sprintf "OVERLAYNTIFF_OPTIONS=\"-c zip -s %s -p %s -b %s -m $MERGEMETHOD",
+    $string .= sprintf "OVERLAYNTIFF_OPTIONS=\"-c zip -s %s -p %s -b %s -m $MERGEMETHOD",
         $pyramid->getImageSpec()->getPixel()->getSamplesPerPixel(),
         $pyramid->getImageSpec()->getPixel()->getPhotometric(),
         $pyramid->getNodata()->getValue();
@@ -448,6 +530,7 @@ sub getScriptInitialization {
 
     $string .= $ONTFUNCTION;
 
+    $string .= "start_line=\$LINENO\n";
     $string .= "\n";
 
     return $string;
